Add latest UbuChk package (features + bug fixes)
So it turns out that the wrath of the damn ForceRun key, formerly
IsUbuntu, is still with us. I'll keep smashing away at it until it's
finally working as intended, for those of you who're not on Ubuntu, but
still want to make use of UbuChk.

It looks like the `--force-run` flag works as intended, but the
associated key does not, so keep that in mind until this functionality
is finally working properly.

On the bright side, UbuChk is really growing!