Add new UbuChk package (major update)
I was tempted to make this my first 'beta' package, but chose against
it in the end. It should be fine, and I'll be watching it like a hawk.
If any issues arise, please do let me know so I can fix them ASAP.

A video talking about some of the changes (towards the end of it) can
be found here: https://youtu.be/z_YS6lDF2aE