Remove 'md5sum' file; pointless if not signed
Since this file itself could be changed -- in the weird and extremely
unlikely event that somebody were able to gain unauthorized access to
this GitHub account or server(s) -- it makes having this file pointless
unless it's signed with something like GPG. I might set that up in the
future, but for now, this file is just inconvenient and pointless.

Hope you understand!

_This commit message will be used on all repositories which contain
thisfile or similar._